+++ src/main/java/webserver/WebServer.java
@@ -1,23 +1,12 @@
 package webserver;
 
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.net.ServerSocket;
 import java.net.Socket;
-import java.util.Set;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.function.Consumer;
-import java.util.function.Function;
 
-import org.reflections.Reflections;
-import org.reflections.scanners.MethodAnnotationsScanner;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import webserver.annotation.Controller;
-import webserver.annotation.RequestMapping;
-import webserver.http.request.MappedRequest;
 
 public class WebServer {
     private static final Logger logger = LoggerFactory.getLogger(WebServer.class);
@@ -25,7 +14,7 @@ public class WebServer {
     private static final int NUMBER_OF_CORES = Runtime.getRuntime().availableProcessors();
 
     public static void main(String args[]) throws Exception {
-        scanRequestMappingAnnotatedMethod();
+        RequestMapper.scanRequestMappingAnnotatedMethod();
 
         int port = 0;
         if (args == null || args.length == 0) {
@@ -47,40 +36,5 @@ public class WebServer {
                 executorService.execute(requestHandler);
             }
         }
-    }
-
-    private static void scanRequestMappingAnnotatedMethod() throws NoSuchMethodException {
-        Reflections reflections = new Reflections("application/controller");
-        Set<Class<?>> controllerAnnotatedClasses = reflections.getTypesAnnotatedWith(Controller.class);
-
-        Method addToRequestMapper = RequestMapper.class.getMethod("add", MappedRequest.class, Method.class);
-
-        controllerAnnotatedClasses.stream()
-            .map(findRequestMappingAnnotatedMethods())
-            .forEach(applyAnnotatedMethodsTo(addToRequestMapper));
-    }
-
-    private static Function<Class<?>, Set<Method>> findRequestMappingAnnotatedMethods() {
-        return aClass ->
-            new Reflections(aClass.getName(), new MethodAnnotationsScanner())
-                .getMethodsAnnotatedWith(RequestMapping.class);
-    }
-
-    private static Consumer<Set<Method>> applyAnnotatedMethodsTo(Method addToRequestMapper) {
-        return classes -> {
-            classes.forEach(addAnnotatedMethodToRequestMapper(addToRequestMapper));
-        };
-    }
-
-    private static Consumer<Method> addAnnotatedMethodToRequestMapper(Method addToRequestMapper) {
-        return annotatedMethod -> {
-            RequestMapping annotation = annotatedMethod.getAnnotation(RequestMapping.class);
-            MappedRequest mappedRequest = new MappedRequest(annotation.method(), annotation.path());
-            try {
-                addToRequestMapper.invoke(null, mappedRequest, annotatedMethod);
-            } catch (IllegalAccessException | InvocationTargetException e) {
-                e.printStackTrace();
-            }
-        };
     }
 }
